% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_suffix}
\alias{add_suffix}
\title{Recursively update all names in a grobTree by adding a suffix to the name}
\usage{
add_suffix(x, suffix)
}
\arguments{
\item{x}{grobTree}

\item{suffix}{string}
}
\value{
new grobTree
}
\description{
If two grobs in a plot have the same name, then it is not guaranteed that
both will be drawn. This function ensures that every grob in a tree has
a unique name.
}
