% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsvg.R
\name{create_new_GeomPointSVG}
\alias{create_new_GeomPointSVG}
\title{Create a fresh instance of a GeomPointSVG ggproto object}
\usage{
create_new_GeomPointSVG()
}
\value{
ggproto object for GeomPointSVG
}
\description{
These Geoms are created dynamically as each time \code{geom_point_svg()}
is called, it wants to customize the \code{$default_aes} on this
ggproto.  Because ggproto objects are environments, then setting a default
on the global copy would set the default for all references. This will
get messy as things like CSS Aesthetics should not be shared between
plots.
}
\details{
So every geom_point_svg() gets a fresh GeomPointSVG by calling this function.
}
