% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-aes.R
\name{css}
\alias{css}
\title{Transform a CSS aesthetic selector into its canonical form for ggsvg}
\usage{
css(selector, ...)
}
\arguments{
\item{selector}{CSS selector as a single character string e.g. "circle .big"}

\item{...}{single named argument of the form \code{css_property = value}.
The \code{value} will remain unevaluated for passing into \code{ggplot2::aes()}}
}
\value{
length-1 named list where the name is the full name of this CSS
        aesthetic, and the value is the unevaluated value passed in to the ...
}
\description{
In general, the user will not want or need to inspect the output of this
funciton.  Instead it is always used within an 'aes()' call for
\code{geom_point_svg()}, or one of the scale functions in \code{ggsvg} of
the form \code{scale_svg_*()}.
}
\examples{
\dontrun{
css("circle .big", stroke = as.factor(cyl))
}

}
