% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-defaults.R
\name{ggplot_add.ScaleSVGDefault}
\alias{ggplot_add.ScaleSVGDefault}
\title{S3 method}
\usage{
\method{ggplot_add}{ScaleSVGDefault}(object, plot, objectname)
}
\arguments{
\item{object, plot, objectname}{see ggplot2 docs}
}
\description{
S3 method
}
