% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-other.R
\name{scale_svg_alpha}
\alias{scale_svg_alpha}
\alias{scale_svg_alpha_binned}
\alias{scale_svg_alpha_continuous}
\alias{scale_svg_alpha_discrete}
\alias{scale_svg_alpha_identity}
\alias{scale_svg_alpha_manual}
\alias{scale_svg_alpha_ordinal}
\alias{scale_svg_colour_binned}
\alias{scale_svg_colour_brewer}
\alias{scale_svg_colour_continuous}
\alias{scale_svg_colour_discrete}
\alias{scale_svg_colour_fermenter}
\alias{scale_svg_colour_grey}
\alias{scale_svg_colour_hue}
\alias{scale_svg_colour_identity}
\alias{scale_svg_colour_manual}
\alias{scale_svg_colour_ordinal}
\alias{scale_svg_colour_qualitative}
\alias{scale_svg_colour_steps}
\alias{scale_svg_colour_steps2}
\alias{scale_svg_colour_stepsn}
\alias{scale_svg_colour_viridis_b}
\alias{scale_svg_colour_viridis_d}
\alias{scale_svg_continuous_identity}
\alias{scale_svg_discrete_identity}
\alias{scale_svg_discrete_manual}
\alias{scale_svg_fill_binned}
\alias{scale_svg_fill_brewer}
\alias{scale_svg_fill_continuous}
\alias{scale_svg_fill_discrete}
\alias{scale_svg_fill_fermenter}
\alias{scale_svg_fill_grey}
\alias{scale_svg_fill_hue}
\alias{scale_svg_fill_identity}
\alias{scale_svg_fill_manual}
\alias{scale_svg_fill_ordinal}
\alias{scale_svg_fill_qualitative}
\alias{scale_svg_fill_steps}
\alias{scale_svg_fill_steps2}
\alias{scale_svg_fill_stepsn}
\alias{scale_svg_fill_viridis_b}
\alias{scale_svg_fill_viridis_d}
\alias{scale_svg_linetype}
\alias{scale_svg_linetype_binned}
\alias{scale_svg_linetype_continuous}
\alias{scale_svg_linetype_discrete}
\alias{scale_svg_linetype_identity}
\alias{scale_svg_linetype_manual}
\alias{scale_svg_radius}
\alias{scale_svg_size}
\alias{scale_svg_size_area}
\alias{scale_svg_size_binned}
\alias{scale_svg_size_binned_area}
\alias{scale_svg_size_continuous}
\alias{scale_svg_size_discrete}
\alias{scale_svg_size_identity}
\alias{scale_svg_size_manual}
\alias{scale_svg_size_ordinal}
\title{Scales for working with arbitrary namedsaestheticcs}
\usage{
scale_svg_alpha(aesthetics, ..., range = c(0.1, 1))

scale_svg_alpha_binned(aesthetics, ..., range = c(0.1, 1))

scale_svg_alpha_continuous(aesthetics, ..., range = c(0.1, 1))

scale_svg_alpha_discrete(aesthetics, ...)

scale_svg_alpha_identity(aesthetics, ..., guide = "none")

scale_svg_alpha_manual(
  aesthetics,
  ...,
  values,
  breaks = waiver(),
  na.value = NA
)

scale_svg_alpha_ordinal(aesthetics, ..., range = c(0.1, 1))

scale_svg_colour_binned(
  aesthetics,
  ...,
  type = getOption("ggplot2.binned.colour")
)

scale_svg_colour_brewer(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = 1
)

scale_svg_colour_continuous(
  aesthetics,
  ...,
  type = getOption("ggplot2.continuous.colour")
)

scale_svg_colour_discrete(
  aesthetics,
  ...,
  type = getOption("ggplot2.discrete.colour")
)

scale_svg_colour_fermenter(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_colour_grey(
  aesthetics,
  ...,
  start = 0.2,
  end = 0.8,
  na.value = "red"
)

scale_svg_colour_hue(
  aesthetics,
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50"
)

scale_svg_colour_identity(aesthetics, ..., guide = "none")

scale_svg_colour_manual(
  aesthetics,
  ...,
  values,
  breaks = waiver(),
  na.value = "grey50"
)

scale_svg_colour_ordinal(
  aesthetics,
  ...,
  type = getOption("ggplot2.ordinal.colour", getOption("ggplot2.ordinal.fill"))
)

scale_svg_colour_qualitative(
  aesthetics,
  ...,
  type = NULL,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50"
)

scale_svg_colour_steps(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_colour_steps2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_colour_stepsn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps",
  colors
)

scale_svg_colour_viridis_b(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_colour_viridis_d(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D"
)

scale_svg_continuous_identity(aesthetics, ..., guide = "none")

scale_svg_discrete_identity(aesthetics, ..., guide = "none")

scale_svg_discrete_manual(aesthetics, ..., values, breaks = waiver())

scale_svg_fill_binned(aesthetics, ..., type = getOption("ggplot2.binned.fill"))

scale_svg_fill_brewer(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = 1
)

scale_svg_fill_continuous(
  aesthetics,
  ...,
  type = getOption("ggplot2.continuous.fill")
)

scale_svg_fill_discrete(
  aesthetics,
  ...,
  type = getOption("ggplot2.discrete.fill")
)

scale_svg_fill_fermenter(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_fill_grey(aesthetics, ..., start = 0.2, end = 0.8, na.value = "red")

scale_svg_fill_hue(
  aesthetics,
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50"
)

scale_svg_fill_identity(aesthetics, ..., guide = "none")

scale_svg_fill_manual(
  aesthetics,
  ...,
  values,
  breaks = waiver(),
  na.value = "grey50"
)

scale_svg_fill_ordinal(
  aesthetics,
  ...,
  type = getOption("ggplot2.ordinal.fill", getOption("ggplot2.ordinal.colour"))
)

scale_svg_fill_qualitative(
  aesthetics,
  ...,
  type = NULL,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50"
)

scale_svg_fill_steps(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_fill_steps2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_fill_stepsn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps",
  colors
)

scale_svg_fill_viridis_b(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps"
)

scale_svg_fill_viridis_d(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D"
)

scale_svg_linetype(aesthetics, ..., na.value = "blank")

scale_svg_linetype_binned(aesthetics, ..., na.value = "blank")

scale_svg_linetype_continuous(aesthetics, ...)

scale_svg_linetype_discrete(aesthetics, ..., na.value = "blank")

scale_svg_linetype_identity(aesthetics, ..., guide = "none")

scale_svg_linetype_manual(
  aesthetics,
  ...,
  values,
  breaks = waiver(),
  na.value = "blank"
)

scale_svg_radius(
  aesthetics,
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)

scale_svg_size(
  aesthetics,
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)

scale_svg_size_area(aesthetics, ..., max_size = 6)

scale_svg_size_binned(
  aesthetics,
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  n.breaks = NULL,
  nice.breaks = TRUE,
  trans = "identity",
  guide = "bins"
)

scale_svg_size_binned_area(aesthetics, ..., max_size = 6)

scale_svg_size_continuous(
  aesthetics,
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)

scale_svg_size_discrete(aesthetics, ...)

scale_svg_size_identity(aesthetics, ..., guide = "none")

scale_svg_size_manual(
  aesthetics,
  ...,
  values,
  breaks = waiver(),
  na.value = NA
)

scale_svg_size_ordinal(aesthetics, ..., range = c(2, 6))
}
\arguments{
\item{aesthetics}{names of aesthethics to apply this scale to. e.g. \code{'fill_rect'}}

\item{..., range, guide, values, breaks, na.value, type, palette}{See \code{ggplot2} documentation}

\item{direction, start, end, h, c, l, h.start, low, high, space}{See \code{ggplot2} documentation}

\item{mid, midpoint, colours, colors, alpha, begin, option, name}{See \code{ggplot2} documentation}

\item{labels, limits, trans, max_size, n.breaks, nice.breaks}{See \code{ggplot2} documentation}
}
\description{
Scales for working with arbitrary namedsaestheticcs
}
