% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-colourbar.R, R/scales.R
\name{scale_svg_colour_gradient}
\alias{scale_svg_colour_gradient}
\alias{scale_svg_colour_distiller}
\alias{scale_svg_colour_gradient2}
\alias{scale_svg_colour_gradientn}
\alias{scale_svg_colour_viridis_c}
\alias{scale_svg_fill_distiller}
\alias{scale_svg_fill_gradient}
\alias{scale_svg_fill_gradient2}
\alias{scale_svg_fill_gradientn}
\alias{scale_svg_fill_viridis_c}
\title{Continuous scales for colour and fill aesthetics for \code{ggsvg}}
\usage{
scale_svg_colour_gradient(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_distiller(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_gradient2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_gradientn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics),
  colors
)

scale_svg_colour_viridis_c(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_distiller(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradient(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradient2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradientn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics),
  colors
)

scale_svg_fill_viridis_c(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_gradient(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_distiller(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_gradient2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_colour_gradientn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics),
  colors
)

scale_svg_colour_viridis_c(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_distiller(
  aesthetics,
  ...,
  type = "seq",
  palette = 1,
  direction = -1,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradient(
  aesthetics,
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradient2(
  aesthetics,
  ...,
  low = muted("red"),
  mid = "white",
  high = muted("blue"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)

scale_svg_fill_gradientn(
  aesthetics,
  ...,
  colours,
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics),
  colors
)

scale_svg_fill_viridis_c(
  aesthetics,
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = ggplot2::guide_colorbar(available_aes = aesthetics)
)
}
\arguments{
\item{aesthetics}{name of the aesthetic e.g. \code{fill_rect}}

\item{..., guide, low, high, space, na.value, type, palette, direction, values, mid, midpoint, colours, colors, alpha, begin, end, option}{see \code{ggplot2} documentation}
}
\description{
All these colour/fill scales use \code{guide_colourbar()} but by default,
this guide will only accept aesthetics of \code{fill} and \code{colour}.

All these colour/fill scales use \code{guide_colourbar()} but by default,
this guide will only accept aesthetics of \code{fill} and \code{colour}.
}
\details{
The two key changes to these scales compared to their \code{ggplot2} originals are:

\itemize{
\item{Default \code{guide} argument is now a \code{guide_colourbar()} object
      which explicitly supports the current \code{aesthetics}}
\item{\code{aesthetics} is now a required argument}
}

The two key changes to these scales compared to their \code{ggplot2} originals are:

\itemize{
\item{Default \code{guide} argument is now a \code{guide_colourbar()} object
      which explicitly supports the current \code{aesthetics}}
\item{\code{aesthetics} is now a required argument}
}
}
