% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvg.R
\name{svg_to_rasterGrob}
\alias{svg_to_rasterGrob}
\title{Convert SVG to a \code{grid} \code{rasterGrob} object}
\usage{
svg_to_rasterGrob(svg_text, width = NULL, height = NULL, css = NULL, ...)
}
\arguments{
\item{svg_text}{character string containing valid SVG}

\item{width, height}{output width,height in pixels or \code{NULL} (the default)
which inputes the size from the SVG}

\item{css}{character string containing CSS text.
This requires your system has a recent version of librsvg.}

\item{...}{other arguments passed to `grid::rasterGrob()`.}
}
\value{
\code{grid::rasterGrob()} object containing the
        \code{nativeRaster} returned from \code{rsvg}
}
\description{
Convert SVG to a \code{grid} \code{rasterGrob} object
}
