% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_suffix}
\alias{add_suffix}
\title{Recursively update all names in a grobTree by adding a suffix}
\usage{
add_suffix(x, suffix)
}
\arguments{
\item{x}{grobTree}

\item{suffix}{string}
}
\value{
new grobTree
}
\description{
Recursively update all names in a grobTree by adding a suffix
}
