% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css-aes.R
\name{css}
\alias{css}
\title{Transform a CSS aesthetic selector into its canonical form for ggsvg}
\usage{
css(selector, ..., format = NULL)
}
\arguments{
\item{selector}{CSS selector as a single character string e.g. "circle .big"}

\item{...}{single named argument of the form \code{css_property = value}.
The \code{value} will remain unevaluated for passing into \code{ggplot2::aes()}}

\item{format}{a string specifying the formatting for the CSS property value.
Default: NULL is equivalent to "[x]" and will insert just the bare value.
For example, if the CSS property required an explicit
"px" suffix on the value, the format would be "[x]px"}
}
\value{
length-1 named list where the name is the full name of this CSS
        aesthetic, and the value is the unevaluated value passed in to the ...
}
\description{
In general, the user will not want or need to inspect the output of this
funciton.  Instead it is always used within an 'aes()' call for
\code{geom_point_svg()}, or one of the scale functions in \code{ggsvg} of
the form \code{scale_svg_*()}.
}
\examples{
\dontrun{
# circle .bit { stroke = XX; }
css("circle .big", stroke = as.factor(cyl))

# circle .bit { stroke = XXpx; }
css("circle .big", stroke = as.factor(cyl), format = "[x]px")
}

}
