% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsvg.R
\name{geom_point_svg}
\alias{geom_point_svg}
\title{Draw SVG}
\usage{
geom_point_svg(
  mapping = NULL,
  ...,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  defaults = list()
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
documentation for \code{ggplot2::geom_point()}}

\item{defaults}{named list of default values for new aesthetics

Set `options(GGSVG_DEBUG = TRUE)` for some verbose debugging which will
cause `{ggsvg}` to output (to the console) the final SVG for each and every element
in the plot.}
}
\value{
ggplot2 layer object
}
\description{
\emph{Aesthetics}
\describe{
\item{svg}{SVG as character string}
\item{svg_width,svg_height}{Specify rendered width and/or height.  If only one of these values
      is specified, then the other will be scaled to keep he aspect ratio.
      If neither value is specified (the default) then dimensions will be taken from the
      SVG itself.  This value could be used to increase the resolution of the SVG so it
      does not appear blurry once rendered to an element in the plot e.g. \code{svg_width = 1000}}
\item{hjust,vjust}{The justification of the SVG's bounding rectangle relative to
      its (x,y) position. Default value of 0.5 mean to centre the SVG at
      the specified location. Standard values for these variables are in the range [0,1]}
\item{x_abs,y_abs}{Absolute positioning within the panel. Default: NULL means
      that the standard x,y positioning is to be used.  Standard values
      for these variables are in the range [0,1]}
}
}
