% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsvg.R
\name{geom_point_svg}
\alias{geom_point_svg}
\title{Draw SVG}
\usage{
geom_point_svg(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  defaults = list()
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
documentation for \code{ggplot2::geom_point()}}

\item{defaults}{named list of default values for new aesthetics}
}
\description{
Draw SVG
}
