% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-defaults.R
\name{scale_ggsvg_default}
\alias{scale_ggsvg_default}
\title{Auto aesthetics for those marked with a ":"}
\usage{
scale_ggsvg_default(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
ggplot object with default scales added to the plot environment
        for any aesthetic which nominates its type
}
\description{
Auto aesthetics for those marked with a ":"
}
