% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvg.R
\name{svg_to_rasterGrob}
\alias{svg_to_rasterGrob}
\title{Convert SVG to rasterGrob via rsvg}
\usage{
svg_to_rasterGrob(svg_text, width = NULL, height = NULL, css = NULL, ...)
}
\arguments{
\item{svg_text}{charcter string containing valid SVG}

\item{width}{output width in pixels or \code{NULL} for default size derived from
SVG size if possible}

\item{height}{output height in pixels or \code{NULL} for default size derived from
SVG size if possible}

\item{css}{see documentation for \code{rsvg::rsvg()}.}

\item{...}{other arguments passed to `grdi::rasterGrob()`. A common extra argument is
\code{interpolate = FALSE}.}
}
\value{
rasterGrob
}
\description{
Note: Waiting on built-in nativeraster export from rsvg. PR submitted 2022-04-14
}
\details{
Note: Calling into the \code{rsvg} package C code is not really allowed by
CRAN, but it works for now to get it released for testing.
}
